Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float number; 
-    scanf("%4f", &number);
-    printf("You entered: %4f\n", number);
+    double number; 
+    scanf("%lf", &number);
+    printf("You entered: %lf\n", number);
     return 0;
 }
